Separated WaitingScreen component from PlayerPage
